+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+## [1.41.0](https://github.com/upstars-global/unity-core-modules/compare/v1.40.0...v1.41.0) (2025-10-21)
+
+### 🚀 Features
+
+* **UN-364:** refactor redeemableCompPointsStore & statusCompPointsStore and cover them with tests  ([#236](https://github.com/upstars-global/unity-core-modules/issues/236))
+ ([462a1b6](https://github.com/upstars-global/unity-core-modules/commit/462a1b61698e2922aa57aaebe496b5a480af88e3))
+
+
+
+    * fix: types
+
+    * refactor: move requests from redeemableCompPointsStore to service
+
+    * refactor: move requests from statusCompPointsStore to service
+
+    * test: redeemableCompPointsStore
+
+    * test: statusCompPointsStore
+
+    * fix: checkHasAvailableCards
+
+    * fix: import
+
 ## [1.40.0](https://github.com/upstars-global/unity-core-modules/compare/v1.39.0...v1.40.0) (2025-10-20)
 
 ### 🚀 Features